fix: separate email identity from provider mocks
