+++ src/lib/magic-link-email.ts
@@ -1,5 +1,5 @@
 import { FACTORY_BRAND } from "@/lib/brand";
-import { emailReplyTo, emailSender } from "@/lib/resend";
+import { emailReplyTo, emailSender } from "@/lib/email-identity";
 import { resolveVerticalConfig } from "@/lib/verticals/registry";
 import type { VerticalId } from "@/lib/verticals/types";
 
@@ -39,7 +39,7 @@ export function buildMagicLinkEmail(input: {
     ? "operator console"
     : multipleWorkspaces
       ? "workspace chooser"
-    : `${input.site!.name} dashboard`;
+      : `${input.site!.name} dashboard`;
   const escapedAccountName = escapeHtml(accountName);
   const actionLabel = input.isSuperadmin
     ? "Open console"
